pyproject.toml
[project]
name = "bndes-despesas-viagens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ckanapi>=4.8",
    "cookiecutter-data-science>=2.2.0",
    "ipykernel>=6.29.5",
    "missingno>=0.5.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
